added reroute

+++ src/utilities/serverApi.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import config from '../config/index';
 import * as Storage from '../utilities/asyncStorage';
-import store from "../utilities/store"
+import store from '../utilities/store';
 const callApi = (
   method,
   reqUrl,
@@ -24,7 +24,16 @@ const callApi = (
     //   });
     // }
 
-    console.log(url, 'data', data, 'method', method, 'token', 'header', headers);
+    console.log(
+      url,
+      'data',
+      data,
+      'method',
+      method,
+      'token',
+      'header',
+      headers
+    );
     let options = {
       method,
       url,
@@ -33,16 +42,21 @@ const callApi = (
     };
     if (method === 'get') delete options['data'];
     axios({ ...options })
-      .then(response =>{console.log("inside callApi response",response); resolve(response)})
+      .then(response => {
+        console.log('inside callApi response', response);
+        resolve(response);
+      })
       .catch(error => {
-         console.log("error inside callApi",error.response)
-         if(error.response.status===401)
-         {
-           console.log("status 401")
-           Storage.remove('token');
-           Storage.remove('user');
-          store.getInstance().getKey("Login").navigate("Login")
-         }
+        //  console.log("error inside callApi",error)
+        if (error.response != undefined && error.response.status === 401) {
+          console.log('status 401');
+          Storage.remove('token');
+          Storage.remove('user');
+          store
+            .getInstance()
+            .getKey('Login')
+            .navigate('Login');
+        }
         reject(error);
       });
   });
